+++ packages/web/src/components/Header/styles.ts
@@ -1,67 +1,73 @@
 import styled from 'styled-components'
 
 export const Container = styled.header`
-  display: flex;
-  flex-direction: column;
-
-  -webkit-box-shadow: 0px 0px 50px 5px rgba(0, 0, 0, 0.25);
-  box-shadow: 0px 0px 50px 5px rgba(0, 0, 0, 0.25);
-
-  background: var(--blue-gray-900);
-`
-export const Content = styled.main`
-  width: 76rem;
-  height: 6.25rem;
-
-  margin: 0 auto;
-  padding: 2rem 3.125rem;
+  height: 11.875rem;
 
   display: flex;
-  align-items: center;
   justify-content: space-between;
 
-  img {
-    width: 12.5rem;
+  border-bottom: 1px solid var(--gray-600);
 
-    cursor: pointer;
-  }
-
-  button {
-    padding: 0.5rem 1.5rem;
+  div {
+    width: 8.75rem;
 
     display: flex;
-    align-items: center;
-    justify-content: center;
+    flex-direction: column;
 
-    gap: 0.5rem;
-    border: 2px solid var(--gray-600);
-    border-radius: 0.5rem;
+    button {
+      padding: 0.5rem 1.5rem;
 
-    background: var(--blue-gray-900);
+      display: flex;
+      align-items: center;
+      justify-content: center;
 
-    svg {
-      height: 25px;
-      width: 25px;
+      gap: 0.5rem;
+      border: 2px solid var(--blue-200);
+      border-radius: 0.5rem;
 
-      color: var(--gray-600);
-
-      transition: color 0.4s;
-    }
-
-    span {
-      font-weight: 500;
-
-      color: var(--white);
-    }
-
-    transition: border-color 0.4s;
-
-    &:hover {
-      border-color: var(--gray-700);
+      background: transparent;
 
       svg {
-        color: var(--gray-700);
+        height: 25px;
+        width: 25px;
+
+        color: var(--blue-200);
+      }
+
+      span {
+        font-weight: 500;
+
+        color: var(--white);
+      }
+
+      transition: filter 0.4s;
+
+      &:hover {
+        filter: brightness(0.7);
       }
     }
+  }
+`
+export const Content = styled.main`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+
+  text-align: center;
+
+  h1 {
+    font-weight: 200;
+    font-size: 1.7rem;
+    letter-spacing: 0.2rem;
+  }
+
+  strong {
+    font-size: 1.7rem;
+  }
+
+  span {
+    font-weight: 200;
+
+    color: var(--blue-200);
   }
 `
